pyproject.toml
[project]
name = "miku"
version = "0.1.0"
description = "Discord leveling bot like Arcane with slash and prefix commands"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.7.1",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "aiohttp>=3.9.0",
]
